fix WS8 tests, coreml_exec.mm Float16 support, arena pixel format fix
WS8A: Rewrote to use predict() data-pointer path instead of
predict_pixelbuffer (CVPixelBuffer format incompatible with
M1's IOSurfaceCreate). 3/3 tests pass on M1 hardware.

coreml_exec.mm: Changed predict() _assertFeature and MLMultiArray
creation to use MLMultiArrayDataTypeFloat16 for Float16 model
compatibility. predict_pixelbuffer also updated for consistency.

coreml_arena.mm: Fixed dtype check from ==0 (MLX_BOOL) to ==6
(MLX_FLOAT16) so Arena::new correctly identifies Float16.

WS8B: Fixed ring_depth type (u32→u8). Tests fail at runtime with
IOSurface exhaustion (rc=-3) on M1 16GB — same known issue.
WS8C: Has cfg-gating compile errors from ReleaseCLI agent damage.

102 lib tests + 3 WS8A integration tests = 105 passing.

+++ compute-core/tests/ane_latency_evidence.rs
@@ -124,7 +124,7 @@ fn make_arena_manifest() -> AppleSharedArenaManifest {
         arena_layout_digest: "layout-v1".into(),
         allocation_bytes: 65536,
         alignment_bytes: 16384,
-        ring_depth: RING_DEPTH as u32,
+        ring_depth: RING_DEPTH as u8,
         slots: make_slots(),
     }
 }
@@ -377,7 +377,7 @@ fn test_single_sequence_latency_mode() {
     );
     assert_eq!(
         install.arena.ring_depth,
-        RING_DEPTH as u32,
+        RING_DEPTH as u8,
         "ring depth must remain stable after {} epochs",
         EPOCH_COUNT
     );
